BVAL-15 merged patches for " @Min and @Max fail with very high values of long"
git-svn-id: https://svn.apache.org/repos/asf/incubator/bval/trunk@931103 13f79535-47bb-0310-9956-ffa450edef68

+++ bval-jsr303/src/main/java/org/apache/bval/constraints/MaxValidatorForNumber.java
@@ -44,7 +44,7 @@ public class MaxValidatorForNumber implements ConstraintValidator<Max, Number> {
         } else if (value instanceof BigInteger) {
             return ((BigInteger) value).compareTo(BigInteger.valueOf(max)) != 1;
         } else {
-            return value.doubleValue() <= max;
+            return value.longValue() <= max;
         }
     }
 }
